+++ tests/test_checks_other.py
@@ -5,7 +5,7 @@ from sqlalchemy.orm import Query
 from threedi_modelchecker.checks.other import BankLevelCheck
 from threedi_modelchecker.checks.other import ConnectionNodesDistance
 from threedi_modelchecker.checks.other import ConnectionNodesLength
-from threedi_modelchecker.checks.other import OpenChannelsWithNestedNewton, StartPointLocationCheck, EndPointLocationCheck, CrossSectionLocationCheck
+from threedi_modelchecker.checks.other import OpenChannelsWithNestedNewton, LinestringLocationCheck, CrossSectionLocationCheck
 from threedi_modelchecker.threedi_model import constants
 from threedi_modelchecker.threedi_model import models
 
@@ -193,13 +193,13 @@ def test_node_distance(session):
 
 @pytest.mark.parametrize("channel_geom", [
     "LINESTRING(155000 463000, 155000 463010)",
-    "LINESTRING(155000 463000, 155000 463999)",  # endpoint is wrong
-    "LINESTRING(155000 463001, 155000 463010)",  # startpoint within tolerance
+    "LINESTRING(155001 463000, 155001 463010)",  # within tolerance
+    "LINESTRING(155000 463010, 155000 463000)",  # reversed
+    "LINESTRING(155001 463010, 155001 463000)",  # reversed, within tolerance
 ])
-def test_channels_start_point_check(session, channel_geom):
+def test_channels_location_check(session, channel_geom):
     if session.bind.name == "postgresql":
         pytest.skip("Postgres only accepts coords in epsg 4326")
-    factories.NumericalSettingsFactory(use_of_nested_newton=0)
     factories.ChannelFactory(
         connection_node_start=factories.ConnectionNodeFactory(
             the_geom="SRID=28992;POINT(155000 463000)"
@@ -210,41 +210,17 @@ def test_channels_start_point_check(session, channel_geom):
         the_geom=f"SRID=28992;{channel_geom}",
     )
 
-    errors = StartPointLocationCheck(column=models.Channel.the_geom, max_distance=1.01).get_invalid(session)
+    errors = LinestringLocationCheck(column=models.Channel.the_geom, max_distance=1.01).get_invalid(session)
     assert len(errors) == 0
 
 
 @pytest.mark.parametrize("channel_geom", [
-    "LINESTRING(155000 463002, 155000 463010)",
-    "LINESTRING(155002 463000, 155000 463010)",
-])
-def test_channels_start_point_check_invalid(session, channel_geom):
-    if session.bind.name == "postgresql":
-        pytest.skip("Postgres only accepts coords in epsg 4326")
-    factories.NumericalSettingsFactory(use_of_nested_newton=0)
-    factories.ChannelFactory(
-        connection_node_start=factories.ConnectionNodeFactory(
-            the_geom="SRID=28992;POINT(155000 463000)"
-        ),
-        connection_node_end=factories.ConnectionNodeFactory(
-            the_geom="SRID=28992;POINT(155000 463010)"
-        ),
-        the_geom=f"SRID=28992;{channel_geom}",
-    )
-
-    errors = StartPointLocationCheck(column=models.Channel.the_geom, max_distance=1.01).get_invalid(session)
-    assert len(errors) == 1
-
-
-@pytest.mark.parametrize("channel_geom", [
-    "LINESTRING(155000 463000, 155000 463010)",
     "LINESTRING(155000 463999, 155000 463010)",  # startpoint is wrong
-    "LINESTRING(155000 463000, 155000 463011)",  # endpoint within tolerance
+    "LINESTRING(155000 463000, 155000 463999)",  # endpoint is wrong
 ])
-def test_channels_end_point_check(session, channel_geom):
+def test_channels_location_check_invalid(session, channel_geom):
     if session.bind.name == "postgresql":
         pytest.skip("Postgres only accepts coords in epsg 4326")
-    factories.NumericalSettingsFactory(use_of_nested_newton=0)
     factories.ChannelFactory(
         connection_node_start=factories.ConnectionNodeFactory(
             the_geom="SRID=28992;POINT(155000 463000)"
@@ -255,29 +231,7 @@ def test_channels_end_point_check(session, channel_geom):
         the_geom=f"SRID=28992;{channel_geom}",
     )
 
-    errors = EndPointLocationCheck(column=models.Channel.the_geom, max_distance=1.01).get_invalid(session)
-    assert len(errors) == 0
-
-
-@pytest.mark.parametrize("channel_geom", [
-    "LINESTRING(155000 463000, 155002 463010)",
-    "LINESTRING(155000 463000, 155000 463012)",
-])
-def test_channels_end_point_check_invalid(session, channel_geom):
-    if session.bind.name == "postgresql":
-        pytest.skip("Postgres only accepts coords in epsg 4326")
-    factories.NumericalSettingsFactory(use_of_nested_newton=0)
-    factories.ChannelFactory(
-        connection_node_start=factories.ConnectionNodeFactory(
-            the_geom="SRID=28992;POINT(155000 463000)"
-        ),
-        connection_node_end=factories.ConnectionNodeFactory(
-            the_geom="SRID=28992;POINT(155000 463010)"
-        ),
-        the_geom=f"SRID=28992;{channel_geom}",
-    )
-
-    errors = EndPointLocationCheck(column=models.Channel.the_geom, max_distance=1.01).get_invalid(session)
+    errors = LinestringLocationCheck(column=models.Channel.the_geom, max_distance=1.01).get_invalid(session)
     assert len(errors) == 1
 
 
@@ -294,15 +248,6 @@ def test_cross_section_location(session):
     factories.CrossSectionLocationFactory(
         channel=channel,
         the_geom=f"SRID=28992;POINT(155001 463008)"
-    )
-
-    # channel with only 1 cross section location; don't check
-    channel2 = factories.ChannelFactory(
-        the_geom=f"SRID=28992;LINESTRING(155010 463000, 155010 463010)",
-    )
-    factories.CrossSectionLocationFactory(
-        channel=channel2,
-        the_geom=f"SRID=28992;POINT(155000 463002)"
     )
     errors = CrossSectionLocationCheck(0.1).get_invalid(session)
     assert len(errors) == 1