+++ threedi_modelchecker/config.py
@@ -27,10 +27,9 @@ from .checks.other import BankLevelCheck
 from .checks.other import BoundaryCondition1DObjectNumberCheck
 from .checks.other import ConnectionNodesDistance
 from .checks.other import ConnectionNodesLength
-from .checks.other import CrossSectionLocationCheckSingle, CrossSectionLocationCheckMultiple
-from .checks.other import EndPointLocationCheck
+from .checks.other import CrossSectionLocationCheck
+from .checks.other import LinestringLocationCheck
 from .checks.other import OpenChannelsWithNestedNewton
-from .checks.other import StartPointLocationCheck
 from .checks.other import Use0DFlowCheck
 from .checks.timeseries import TimeseriesIncreasingCheck
 from .checks.timeseries import TimeseriesRowCheck
@@ -206,8 +205,9 @@ CHECKS += [
 ## 005x: CROSS SECTIONS
 
 CHECKS += [
-    CrossSectionLocationCheckSingle(level=CheckLevel.WARNING, max_distance=1.0, error_code=52),
-    CrossSectionLocationCheckMultiple(max_distance=1.0, error_code=52),
+    CrossSectionLocationCheck(
+        level=CheckLevel.WARNING, max_distance=1.0, error_code=52
+    ),
     OpenChannelsWithNestedNewton(error_code=53),
     QueryCheck(
         error_code=54,
@@ -517,11 +517,7 @@ CHECKS += [
     for table in [models.Orifice, models.Weir]
 ]
 CHECKS += [
-    StartPointLocationCheck(error_code=205, column=table.the_geom, max_distance=0.1)
-    for table in [models.Channel, models.Culvert]
-]
-CHECKS += [
-    EndPointLocationCheck(error_code=206, column=table.the_geom, max_distance=0.1)
+    LinestringLocationCheck(error_code=205, column=table.the_geom, max_distance=1)
     for table in [models.Channel, models.Culvert]
 ]